requirements.txt
python-dotenv
ebird-api
fastmcp
